Rename component to be aligned with naming convention
