+++ game-frontend/src/components/visual/CanvasWrapper.tsx
